requirements.txt
anytree
aiofiles
appdirs
aria2p
asyncio
attrdict
beautifulsoup4
cfscrape
cloudscraper
feedparser
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
hachoir
html_telegraph_poster
lk21
lxml
megasdkrestclient
pillow
psutil
psycopg2-binary
pybase64
pyrogram
python-dotenv
pykeyboard
python-magic
pycountry
python-telegram-bot==13.15
requests
telegraph
tenacity
tgCrypto
urllib3
yt-dlp
vcsi
speedtest-cli
zippyshare_downloader
